update pointwise_dynamic for cpp wrapper compat
